+++ explicitly-unused.js
@@ -14,7 +14,6 @@ module.exports = {
     "init-declarations": ["off"],
     "line-comment-position": ["off"],
     "lines-around-comment": ["off"],
-    "lines-between-class-members": ["off"],
     "max-classes-per-file": ["off"],
     "max-depth": ["off"],
     "max-lines-per-function": ["off"],